+++ compose_rl/algorithms/online/callback.py
@@ -1,7 +1,7 @@
 # Copyright 2024 MosaicML ComposeRL authors
 # SPDX-License-Identifier: Apache-2.0
 
-"""PPO callback."""
+"""Online On-Policy RL callback."""
 
 from __future__ import annotations
 
@@ -15,6 +15,8 @@ from typing import Any, Optional, Union
 import ray
 import torch
 import wandb
+from transformers import PreTrainedTokenizer, PreTrainedTokenizerFast
+
 from composer.core import (
     Precision,
     State,
@@ -27,17 +29,15 @@ from composer.loggers import Logger, MLFlowLogger, WandBLogger
 from composer.trainer.trainer import _get_initial_device_train_microbatch_size
 from composer.utils import dist, ensure_tuple
 from llmfoundry.interfaces import CallbackWithConfig
-from transformers import PreTrainedTokenizer, PreTrainedTokenizerFast
 
-import compose_rl.utils as utils
-from compose_rl.ppo.buffer import MinibatchRolloutBuffer
-from compose_rl.ppo.generation_utils import hf_generate, vllm_generate
-from compose_rl.ppo.model import ComposerHFPolicyModel, ComposerMosaicPolicy
-from compose_rl.ppo.reward_manager import (
+from compose_rl.algorithms.online.generation_utils import hf_generate, vllm_generate
+from compose_rl.algorithms.online.model import ComposerHFPolicyModel, ComposerMosaicPolicy
+from compose_rl.algorithms.online.reward_manager import (
     ReferenceOutput,
     RewardManager,
     RewardOutput,
 )
+from compose_rl.data.buffer import MinibatchRolloutBuffer
 from compose_rl.registry_builders import build_kl_controller
 from compose_rl.utils import (
     add_right_padding,
@@ -45,19 +45,21 @@ from compose_rl.utils import (
     compute_advantages,
     create_vllm_engines,
     dist_compute_masked_mean_and_var,
+    flatten,
     get_decoded_sequence,
     get_entropies,
     get_log_probs,
     init_process_group,
     mask_eos,
     masked_mean,
+    masked_sum,
     switch_left_to_right_padding,
 )
 
 Tokenizer = Union[PreTrainedTokenizer, PreTrainedTokenizerFast]
 Policy = Union[ComposerHFPolicyModel, ComposerMosaicPolicy]
 
-__all__ = ['PPOCallback', 'env_reward']
+__all__ = ['OnPolicyCallback', 'env_reward']
 
 log = logging.getLogger(__name__)
 
@@ -303,8 +305,8 @@ def env_reward(
     )
 
 
-class PPOCallback(CallbackWithConfig):
-    """Callback for managing PPO training in an RLHF loop.
+class OnPolicyCallback(CallbackWithConfig):
+    """Callback for managing on-policy training in an RLHF loop.
 
     Args:
         train_config (dict): Training config passed to callback via foundry train.py as
@@ -666,7 +668,7 @@ class PPOCallback(CallbackWithConfig):
                 ret_batch[key] = torch.cat(curr_values)
             else:
                 if key == 'verified_answer':
-                    ret_batch[key] = list(utils.flatten(curr_values))
+                    ret_batch[key] = list(flatten(curr_values))
                 else:
                     # this is an edge case that we will not hit currently, but just handling it as needed
                     ret_batch[key] = curr_values
@@ -865,7 +867,7 @@ class PPOCallback(CallbackWithConfig):
             rewards = env_outs['rewards']
 
             # Flatten the rewards by summing on sequence length/action_mask
-            flat_rewards = utils.masked_sum(
+            flat_rewards = masked_sum(
                 rewards,
                 env_outs['action_mask'],
                 dim=-1,
