+++ compose_rl/data/__init__.py
@@ -1,6 +1,10 @@
 # Copyright 2024 MosaicML ComposeRL authors
 # SPDX-License-Identifier: Apache-2.0
 
+from compose_rl.data.buffer import (
+    DummyDataset,
+    MinibatchRolloutBuffer,
+)
 from compose_rl.data.dataloader import (
     build_finegrained_preference_dataloader,
     build_pairwise_preference_dataloader,
@@ -11,30 +15,14 @@ from compose_rl.data.preference_data import (
     pairwise_preference_dataset_collate_fn,
 )
 from compose_rl.data.prompt_data import prompt_dataset_collate_fn
-from compose_rl.data.rlvr_utils import (
-    extract_gsm8k_answer,
-    extract_math_answer,
-    is_equiv,
-    last_boxed_only_string,
-    normalize_final_answer,
-    prepare_gsm8k_prompt,
-    prepare_math_prompt,
-    remove_boxed,
-)
 
 __all__ = [
     'build_pairwise_preference_dataloader',
     'build_finegrained_preference_dataloader',
     'build_prompt_dataloader',
-    'extract_gsm8k_answer',
+    'DummyDataset',
     'finegrained_preference_dataset_collate_fn',
+    'MinibatchRolloutBuffer',
     'pairwise_preference_dataset_collate_fn',
-    'prepare_gsm8k_prompt',
     'prompt_dataset_collate_fn',
-    'extract_math_answer',
-    'prepare_math_prompt',
-    'last_boxed_only_string',
-    'remove_boxed',
-    'is_equiv',
-    'normalize_final_answer',
 ]
