--- compose_rl/dpo/__init__.py
@@ -1,11 +1,0 @@
-# Copyright 2024 MosaicML ComposeRL authors
-# SPDX-License-Identifier: Apache-2.0
-
-from compose_rl.dpo.callback import DPOCallback
-from compose_rl.dpo.model import ComposerDPOLM, ComposerHFDPOLM
-
-__all__ = [
-    'ComposerDPOLM',
-    'ComposerHFDPOLM',
-    'DPOCallback',
-]
